Rename OrdersUploader to FileUpload in preparation for one true uploader
